+++ reduct/http.py
@@ -14,7 +14,7 @@ API_PREFIX = "/api/v1"
 class HttpClient:
     """Wrapper for HTTP calls"""
 
-    FILE_SIZE_FOR_100_CONTINUE = 256_000
+    FILE_SIZE_FOR_100_CONTINUE = 64_000_000
 
     def __init__(
         self,
@@ -35,6 +35,7 @@ class HttpClient:
         self._timeout = ClientTimeout(timeout)
         self._api_version = None
         self._session = kwargs.pop("session", None)
+        self._verify_ssl = kwargs.pop("verify_ssl", True)
 
     @asynccontextmanager
     async def request(
@@ -63,6 +64,8 @@ class HttpClient:
                 for name, value in kwargs["labels"].items():
                     extra_headers[f"x-reduct-label-{name}"] = str(value)
             del kwargs["labels"]
+
+        kwargs["verify_ssl"] = self._verify_ssl
 
         if self._session is None:
             connector = aiohttp.TCPConnector(force_close=True)
@@ -107,10 +110,8 @@ class HttpClient:
                             response.headers["x-reduct-error"],
                         )
                     raise ReductError(response.status, "Unknown error")
-        except ClientConnectorError:
-            raise ReductError(
-                599, f"Connection failed, server {self._url} cannot be reached"
-            ) from None
+        except ClientConnectorError as error:
+            raise ReductError(599, str(error)) from None
 
     async def request_all(
         self, method: str, path: str = "", **kwargs